requirements.txt
appdirs==1.4.3
click==6.7
decorator==4.0.11
Django==1.10.6
django-appconf==1.0.2
django-compressor==2.1.1
django-libs==1.67.12
djangorestframework==3.6.2
geocoder==1.20.1
Markdown==2.6.8
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
ratelim==0.1.6
rcssmin==1.0.6
requests==2.13.0
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.3
ua-parser==0.7.3
user-agents==1.1.0
whitenoise==3.2.2
